src/requirements.txt
gunicorn==19.10.0
flask==2.0.3
# Bootstrap-Flask==1.2.0
flask-bootstrap==3.3.7.1
requests==2.22.0
google-search-results==2.4.1
web3==5.28.0
pandas
numpy==1.22.2
torch==1.10.2
torchvision==0.11.3
scipy==1.8.0
pillow==9.0.1